# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Benin Tourisme API",
        "description": "API de gestion et promotion des sites touristiques et événements culturels du Bénin",
        "contact": {
            "email": "eldomoreogbohouili@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur API local"
        }
    ],
    "paths": {
...
        "/api/prix": {
            "get": {
                "tags": [
                    "Prix"
                ],
                "summary": "Lister les prix",
                "operationId": "fee5da546673b4198261bf8e6cd8c981",
                "parameters": [
                    {
                        "name": "id_site",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_evnmt",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des prix"
                    }
                }
            },
            "post": {
                "tags": [
                    "Prix"
                ],
                "summary": "Créer un prix",
                "operationId": "426255c3e1e411dae88a3581fb1497af",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle",
                                    "montant"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "montant": {
                                        "type": "number"
                                    },
                                    "id_site": {
                                        "type": "integer"
                                    },
                                    "id_evnmt": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Prix créé"
                    }
                }
            }
        },
        "/api/prix/{id}": {
            "get": {
                "tags": [
                    "Prix"
                ],
                "summary": "Afficher un prix",
                "operationId": "bf5b97aa3b33843ea3e80c67c2226fab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du prix"
                    }
                }
            },
            "put": {
                "tags": [
                    "Prix"
                ],
                "summary": "Mettre à jour un prix",
                "operationId": "db3b0435e0d1f758ccdf542d1e14e55d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "montant": {
                                        "type": "number"
                                    },
                                    "id_site": {
                                        "type": "integer"
                                    },
                                    "id_evnmt": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Prix mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Prix"
                ],
                "summary": "Supprimer un prix",
                "operationId": "8f8f5a30ecb435fdffca8866b405a8bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prix supprimé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Authorization header using the Bearer scheme"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}